Default CMAKE_CXX_STANDARD to 20
So far, the user was required to set the C++ version. It is required
that the user is able to select the version, as some users might want
to work with the C++17 coroutine-ts and others want to work with
C++20.
However, this led to frustration with many users, as just invoking
cmake did not lead to a working build. This patch changes the default
to C++20, informing the user about this choice.